requirements.txt
flask
phonenumbers
timezonefinder
geopy
pillow
pytz